Update README.md
Added the following text: 
The point of the script was to create a simple, light-weight way to quickly start up OpenVPN and connect to a VPN server. My examples are built around config files from a specific commercial VPN service, but as long as you place config files in a location the script can get to, you can use any commercial service or your own personal VPN server(s).